// src/api/models/login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LoginRequest
 */
export interface LoginRequest {
    /**
     * 이메일(아이디)
     * @type {string}
     * @memberof LoginRequest
     */
    'email': string;
    /**
     * 비밀번호
     * @type {string}
     * @memberof LoginRequest
     */
    'password': string;
}