// src/api/models/signup-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SignupRequest
 */
export interface SignupRequest {
    /**
     * 사용자 이메일 (아이디로 사용)
     * @type {string}
     * @memberof SignupRequest
     */
    'email': string;
    /**
     * 비밀번호 
     * @type {string}
     * @memberof SignupRequest
     */
    'password': string;
    /**
     * 사용자 닉네임 (중복 불가)
     * @type {string}
     * @memberof SignupRequest
     */
    'nickname': string;
}