// src/api/models/kakao-login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface KakaoLoginRequest
 */
export interface KakaoLoginRequest {
    /**
     * 카카오에서 발급받은 액세스 토큰
     * @type {string}
     * @memberof KakaoLoginRequest
     */
    'accessToken': string;
}