// src/api/models/naver-login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface NaverLoginRequest
 */
export interface NaverLoginRequest {
    /**
     * 네이버에서 발급받은 액세스 토큰
     * @type {string}
     * @memberof NaverLoginRequest
     */
    'accessToken': string;
}