// src/api/models/auth-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AuthResponse
 */
export interface AuthResponse {
    /**
     * 사용자 식별 번호
     * @type {number}
     * @memberof AuthResponse
     */
    'userId': number;
    /**
     * API 접근에 사용되는 JWT 액세스 토큰
     * @type {string}
     * @memberof AuthResponse
     */
    'accessToken': string;
    /**
     * AccessToken 재발급에 사용되는 JWT 리프레시 토큰
     * @type {string}
     * @memberof AuthResponse
     */
    'refreshToken': string;
    /**
     * 로그인한 사용자의 이메일
     * @type {string}
     * @memberof AuthResponse
     */
    'email': string;
    /**
     * 로그인한 사용자의 닉네임
     * @type {string}
     * @memberof AuthResponse
     */
    'nickname': string;
    /**
     * 로그인 방법
     * @type {string}
     * @memberof AuthResponse
     */
    'provider': AuthResponseProviderEnum;
    /**
     * 사용자 권한
     * @type {string}
     * @memberof AuthResponse
     */
    'role': AuthResponseRoleEnum;
}

export const AuthResponseProviderEnum = {
    Local: 'LOCAL',
    Google: 'GOOGLE',
    Kakao: 'KAKAO',
    Naver: 'NAVER'
} as const;

export type AuthResponseProviderEnum = typeof AuthResponseProviderEnum[keyof typeof AuthResponseProviderEnum];
export const AuthResponseRoleEnum = {
    User: 'USER',
    Admin: 'ADMIN'
} as const;

export type AuthResponseRoleEnum = typeof AuthResponseRoleEnum[keyof typeof AuthResponseRoleEnum];